requirements.txt
torch
transformers
numpy
scikit-learn
python-dotenv
numpy
pandas